+++ src/routes/governance/dreps/drep-id/votes.ts
@@ -5,6 +5,8 @@ import { getDbSync } from '../../../../utils/database.js';
 import { SQLQuery } from '../../../../sql/index.js';
 import { getSchemaForEndpoint } from '@blockfrost/openapi';
 import { isUnpaged } from '../../../../utils/routes.js';
+import { handle400Custom } from '@blockfrost/blockfrost-utils/lib/fastify.js';
+import { drepIdToRaw } from '../../../../utils/bech32.js';
 
 async function route(fastify: FastifyInstance) {
   fastify.route({
@@ -13,6 +15,14 @@ async function route(fastify: FastifyInstance) {
     schema: getSchemaForEndpoint('/governance/dreps/{drep_id}/votes'),
 
     handler: async (request: FastifyRequest<QueryTypes.RequestParametersDRepID>, reply) => {
+      let drepHex;
+
+      try {
+        drepHex = drepIdToRaw(request.params.drep_id);
+      } catch {
+        return handle400Custom(reply, 'Invalid or malformed drep id.');
+      }
+
       const clientDbSync = await getDbSync(fastify);
 
       try {
@@ -24,12 +34,7 @@ async function route(fastify: FastifyInstance) {
             )
           : await clientDbSync.query<QueryTypes.DRepsDrepIDUpdates>(
               SQLQuery.get('governance_dreps_drep_id_votes'),
-              [
-                request.query.order,
-                request.query.count,
-                request.query.page,
-                request.params.drep_id,
-              ],
+              [request.query.order, request.query.count, request.query.page, drepHex],
             );
 
         clientDbSync.release();