+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Querying Dreps with `drep_script` bech32 prefix
+
 ## [2.2.0] - 2024-09-12
 
 ### Added